requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
pytest>=7.0.0
openai>=1.0.0